+++ ls-plugin.php
@@ -44,6 +44,6 @@ add_action( 'init', 'ls_plugin_load_textdomain' );
  * Add include files to inc/ and require them here when ready.
  */
 function ls_plugin_init() {
-	// Example: require_once LS_PLUGIN_PLUGIN_DIR . 'inc/class-ls-plugin-example.php';
+	require_once LS_PLUGIN_PLUGIN_DIR . 'inc/linkable-blocks.php';
 }
 add_action( 'plugins_loaded', 'ls_plugin_init' );